Fixes #1354
Fixes #1354. It is to fix the compiling issue for gcc12.0.1.
[-Werror=address] [-Werror=dangling-pointer=][-Werror=alloc-size-larger-than=]

+++ media_softlet/agnostic/common/os/mos_utilities_next.h
@@ -90,10 +90,10 @@ public:
     template<class _Ty, class... _Types>
     static _Ty *MosNewArrayUtil(const char *functionName,
         const char *filename,
-        int32_t line, int32_t numElements)
+        int32_t line, size_t numElements)
 #else
     template<class _Ty, class... _Types>
-    static _Ty* MosNewArrayUtil(int32_t numElements)
+    static _Ty* MosNewArrayUtil(size_t numElements)
 #endif
     {
 #if (_DEBUG || _RELEASE_INTERNAL)
